example(RecordLoader): Add support for a unique parameter
This will allow it to act as a more generic foreign key loader
so that it can be used to load a `has_many` association. This can
be useful when working with a large association that can't be
preloaded, since the `where` parameter can be used to provide an
appropriate scope condition.