Clear loader cache in Mutation#after_resolve
The `resolve` method of a Mutation can make changes to the DB that
invalidate any cached Loader results. So, the cache should be cleared
after the resolve method has completed, and before any nested fields on
the Mutation query are resolved.

In our codebase, this causes authorization errors if the Mutation
removes a member from a list/connection, and the client requests that
same list field as part of the mutation response. Since the Loader
cached the pre-mutation result, the wrong data is returned to the
client.